leth

+++ 201610_MKT/fb_audit_content_video/rename_downloaded_video.mp.py
@@ -64,8 +64,9 @@ def do_work(list_index,in_queue, out_list):
                     subprocess.call(["mv", file_name,new_file_name])
                 else:
                     print('file exist',new_file_name)
-                    print("/bin/rm -f ", file_name)
-                    subprocess.call(["/bin/rm -f", file_name])
+                    if os.path.exists(file_name):
+                        print("/bin/rm -f ", file_name)
+                        subprocess.call(["/bin/rm -f ", file_name])
 
 
         result = (line_no, video)
